requirements.txt
absl-py==1.2.0
albumentations==1.2.0
blis==0.9.1
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.1
contourpy==1.0.5
cycler==0.11.0
cymem==2.0.6
distlib==0.3.6
efficientnet-pytorch==0.7.1
fastai==2.7.9
fastcore==1.5.27
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.8.0
fonttools==4.37.3
google-auth==2.11.1
google-auth-oauthlib==0.4.6
grpcio==1.49.1
idna==3.4
imageio==2.22.0
importlib-metadata==4.12.0
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
munch==2.5.0
murmurhash==1.0.8
networkx==2.8.6
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.0
pathy==0.6.2
Pillow==9.2.0
platformdirs==2.5.2
preshed==3.0.7
pretrainedmodels==0.7.4
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.3
scikit-learn==1.1.2
scipy==1.9.1
segmentation-models-pytorch==0.3.0
six==1.16.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.1.1
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.4.12
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
wasabi==0.10.1
Werkzeug==2.2.2
zipp==3.8.1
